refactor: create folder for postService
